src/main/java/com/zberg/esc/tippspiel/security/TippspielUserDetailsService.java
package com.zberg.esc.tippspiel.security;

import com.zberg.esc.tippspiel.security.entities.User;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.security.core.userdetails.UserDetails;
import org.springframework.security.core.userdetails.UserDetailsService;
import org.springframework.security.core.userdetails.UsernameNotFoundException;
import org.springframework.stereotype.Service;

@Service
public class TippspielUserDetailsService implements UserDetailsService {

    private final UserRepository userRepository;

    @Autowired
    public TippspielUserDetailsService(final UserRepository userRepository) {
        this.userRepository = userRepository;
    }

    @Override
    public UserDetails loadUserByUsername(String username) throws UsernameNotFoundException {
        final User user = userRepository.findByUsername(username);
        if (user == null) {
            throw new UsernameNotFoundException(username);
        }
        return new TippspielPrincipal(user);
    }
}
